moving to float SYMMATRIX type and fixing no-compile problems

+++ Tensor/itkPreservationOfPrincipalDirectionTensorReorientationImageFilter.cxx
@@ -163,10 +163,10 @@ PreservationOfPrincipalDirectionTensorReorientationImageFilter<TTensorImage, TVe
 }
 
 template <typename TTensorImage, typename TVectorImage>
-itk::SymmetricSecondRankTensor<double, 3>
+itk::SymmetricSecondRankTensor<float, 3>
 PreservationOfPrincipalDirectionTensorReorientationImageFilter<TTensorImage, TVectorImage>
-::ReorientTensors( itk::SymmetricSecondRankTensor<double, 3> fixedTens, itk::SymmetricSecondRankTensor<double,
-                                                                                                       3> movingTens,
+::ReorientTensors( itk::SymmetricSecondRankTensor<float, 3> fixedTens, itk::SymmetricSecondRankTensor<float,
+                                                                                                      3> movingTens,
                    typename TTensorImage::IndexType index)
 {
 // get from and to vectors
